requirements.txt
fastapi[all]
uvicorn == 0.30.1
psycopg2 == 2.9.9
python-dotenv == 1.0.1
starlette == 0.37.2
apscheduler == 3.10.4